fetch/api/request/request-consume-empty.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8">
    <title>Request consume empty bodies</title>
    <meta name="help" href="https://fetch.spec.whatwg.org/#request">
    <meta name="help" href="https://fetch.spec.whatwg.org/#body-mixin">
    <meta name="author" title="Canon Research France" href="https://www.crf.canon.fr">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
    function checkBodyText(test, request) {
      return request.text().then(function(bodyAsText) {
        assert_equals(bodyAsText, "", "Resolved value should be empty");
      });
    }

    function checkBodyBlob(test, request) {
      return request.blob().then(function(bodyAsBlob) {
        var promise = new Promise(function(resolve, reject) {
          var reader = new FileReader();
          reader.onload = function(evt) {
            resolve(reader.result)
          };
          reader.onerror = function() {
            reject("Blob's reader failed");
          };
          reader.readAsText(bodyAsBlob);
        });
        return promise.then(function(body) {
          assert_equals(body, "", "Resolved value should be empty");
        });
      });
    }

    function checkBodyArrayBuffer(test, request) {
      return request.arrayBuffer().then(function(bodyAsArrayBuffer) {
        assert_equals(bodyAsArrayBuffer.byteLength, 0, "Resolved value should be empty");
      });
    }

    function checkBodyJSON(test, request) {
      return request.json().then(
        function(bodyAsJSON) {
          assert_unreached("JSON parsing should fail");
        },
        function() {
        });
    }

    function checkBodyFormData(test, request) {
      return request.formData().then(function(bodyAsFormData) {
        assert_true(bodyAsFormData instanceof FormData, "Should receive a FormData");
     });
    }

    function checkBodyFormDataError(test, request) {
      return promise_rejects(test, new TypeError(), request.formData());
    }

    function checkBodyFormDataText(test, request) {
      return request.text().then(function(bodyAsText) {
        assert_true(request.headers.has("Content-Type"), "Response contains Content-Type header");
        var boundaryMatches = request.headers.get("Content-Type").match(/;\s*boundary=("?)([^";\s]*)\1/);
        assert_true(!!boundaryMatches, "Response contains boundary parameter");
        var boundary = boundaryMatches[2];
        assert_equals(bodyAsText, "--" + boundary + "--\r\n");
     });
    }

    function checkRequestWithNoBody(bodyType, checkFunction, headers = []) {
      promise_test(function(test) {
        var request = new Request("", {"method": "POST", "headers": headers});
        assert_false(request.bodyUsed);
        return checkFunction(test, request).then(function() {
          assert_false(request.bodyUsed);
        });
      }, "Consume request's body as " + bodyType);
    }

    checkRequestWithNoBody("text", checkBodyText);
    checkRequestWithNoBody("blob", checkBodyBlob);
    checkRequestWithNoBody("arrayBuffer", checkBodyArrayBuffer);
    checkRequestWithNoBody("json (error case)", checkBodyJSON);
    checkRequestWithNoBody("formData with correct multipart type (error case)", checkBodyFormDataError, [["Content-Type", 'multipart/form-data; boundary="boundary"']]);
    checkRequestWithNoBody("formData with correct urlencoded type", checkBodyFormData, [["Content-Type", "application/x-www-form-urlencoded;charset=UTF-8"]]);
    checkRequestWithNoBody("formData without correct type (error case)", checkBodyFormDataError);

    function checkRequestWithEmptyBody(bodyType, body, checkFunction) {
      promise_test(function(test) {
        var request = new Request("", {"method": "POST", "body": body});
        assert_false(request.bodyUsed, "bodyUsed is false at init");
        return checkFunction(test, request).then(function() {
          assert_true(request.bodyUsed, "bodyUsed is true after being consumed");
        });
      }, "Consume empty " + bodyType + " request body as " + (checkFunction == checkBodyArrayBuffer ? "arrayBuffer" : "text"));
    }

    // FIXME: Add BufferSource.
    checkRequestWithEmptyBody("blob", new Blob([], { "type" : "text/plain" }), checkBodyArrayBuffer);
    checkRequestWithEmptyBody("text", "", checkBodyArrayBuffer);
    checkRequestWithEmptyBody("blob", new Blob([], { "type" : "text/plain" }), checkBodyText);
    checkRequestWithEmptyBody("text", "", checkBodyText);
    checkRequestWithEmptyBody("URLSearchParams", new URLSearchParams(""), checkBodyText);
    checkRequestWithEmptyBody("FormData", new FormData(), checkBodyFormDataText);
    checkRequestWithEmptyBody("ArrayBuffer", new ArrayBuffer(), checkBodyText);
    </script>
  </body>
</html>
